// src/models/ServiceProvider.ts
export interface ServiceProvider {
  email: string ;
  first_name: string ;
  last_name: string ;
  password: string ;
  currentPassword: string ;
  phone: string ;
  user_type: string ;
  joining_date?: string ;
  active: number ;
  profile?: string ;
  category_id: string ;
  name_ar: string ;
  delivery_charge: string;
  name_en: string ;
  description_ar: string ;
  description_en: string ;
  start_time: string ;
  end_time: string ;
  lat: string;
  lng: string;
  location: string;

  busy: string ;
}